add general transforms that work with spread
