+++ packages/codemods/src/codeMods/utilities/transforms.ts
@@ -1,32 +1,77 @@
-import { PropTransform, EnumMap } from '../types';
-import { JsxExpression, SyntaxKind } from 'ts-morph';
+import { PropTransform, ValueMap } from '../types';
+import { JsxExpression, SyntaxKind, JsxOpeningElement, JsxSelfClosingElement } from 'ts-morph';
 import { Maybe } from '../../maybe';
+import { renamePropInSpread } from './helpers/propHelpers';
 
 // Used in the renameProp utility.
 
-/* Transform function that replaces a JsxExpression node
-   with data specified by the user -- in this case a desired boolean value. */
-export function boolTransform(newValue: boolean): PropTransform {
-  return (element: JsxExpression) => {
-    const exp = Maybe(element.getFirstChildByKind(SyntaxKind.TrueKeyword));
-    if (exp.just) {
-      exp.value.replaceWithText(newValue.toString());
+/* Transform function that returns the new boolean value
+   specified by a user. Less complex than other transforms
+   because the user can simply overwrite data instead of
+   mapping it. If both NEWVALUE and MAP are provided
+   (which should not happen), MAP will be used. If neither
+   are provided, nothing will happen. */
+export function boolTransform(newValue?: boolean, map?: ValueMap<string>): PropTransform {
+  return (
+    element: JsxExpression | JsxOpeningElement | JsxSelfClosingElement,
+    toRename: string,
+    replacementName: string,
+  ) => {
+    if (newValue === undefined && !map) {
+      return;
+    }
+    switch (element.getKind()) {
+      case SyntaxKind.JsxExpression:
+        {
+          const toChange = Maybe(element.getFirstChildByKind(SyntaxKind.TrueKeyword));
+          if (toChange.just) {
+            const oldText = toChange.value.getText();
+            toChange.value.replaceWithText(map ? map[oldText] : newValue!.toString());
+          }
+        }
+        break;
+      case SyntaxKind.JsxOpeningElement:
+      case SyntaxKind.JsxSelfClosingElement: {
+        renamePropInSpread(
+          element as JsxOpeningElement | JsxSelfClosingElement,
+          toRename,
+          replacementName,
+          map,
+          newValue?.toString(),
+        );
+        break;
+      }
     }
   };
 }
 
-/* Transform function that takes in a predefined mapping of old to new
-   enum values. If the enum is in a spread operator, you will need to
-   manually pass the enum map into the renameProp utility function. */
-export function enumTransform(changeValueMap: EnumMap<string>): PropTransform {
-  return (element: JsxExpression) => {
-    const enumExp = Maybe(element.getFirstChildByKind(SyntaxKind.PropertyAccessExpression));
-    const newEnumName = enumExp.then(value => value.getText()).then(value => changeValueMap[value]);
-    const firstChild = enumExp.then(value => value.getFirstChildByKind(SyntaxKind.Identifier));
-    const lastChild = enumExp.then(value => value.getLastChildByKind(SyntaxKind.Identifier));
-    if (firstChild.just && lastChild.just && newEnumName.just) {
-      firstChild.value.replaceWithText(newEnumName.value.substring(0, newEnumName.value.indexOf('.')));
-      lastChild.value.replaceWithText(newEnumName.value.substring(newEnumName.value.indexOf('.') + 1));
+/* Transform function used if the value to change is an ENUM. Follows
+   the same contraints as the above function, except the parameter MAP
+   is not optional. */
+export function enumTransform(map: ValueMap<string>): PropTransform {
+  return (
+    element: JsxExpression | JsxOpeningElement | JsxSelfClosingElement,
+    toRename: string,
+    replacementName: string,
+  ) => {
+    if (!map) {
+      return;
+    }
+    switch (element.getKind()) {
+      case SyntaxKind.JsxExpression:
+        {
+          const toChange = Maybe(element.getFirstChildByKind(SyntaxKind.PropertyAccessExpression));
+          if (toChange.just) {
+            const oldText = toChange.value.getText();
+            toChange.value.replaceWithText(map[oldText]);
+          }
+        }
+        break;
+      case SyntaxKind.JsxOpeningElement:
+      case SyntaxKind.JsxSelfClosingElement: {
+        renamePropInSpread(element as JsxOpeningElement | JsxSelfClosingElement, toRename, replacementName, map);
+        break;
+      }
     }
   };
 }
